disable generate_column

+++ tests/integration_tests/generate_column/run.sh
@@ -9,6 +9,11 @@ CDC_BINARY=cdc.test
 SINK_TYPE=$1
 
 function run() {
+	# storage is not supported yet.
+	if [ "$SINK_TYPE" == "storage" ]; then
+		return
+	fi
+
 	rm -rf $WORK_DIR && mkdir -p $WORK_DIR
 
 	start_tidb_cluster --workdir $WORK_DIR
